Final file split
