+++ config/splat.us.stcen.yaml
@@ -50,6 +50,7 @@ segments:
       - [0xD5C4, .rodata, 18084] # EntityStageNamePopup
       - [0xD5F8, .rodata, prim_helpers]
       - [0xD600, c]
+      - [0xDF0C, c]
       - [0xF890, c]
       - [0x10E4C, c, random]
       - [0x10E7C, c, update]
